add task to check for inactive accounts for long time

+++ backend/src/services/category.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import mongoose, { Model } from 'mongoose';
 import { Category, CategoryDocument } from 'src/schemas/category.schema';
@@ -12,10 +12,10 @@ export class CategoryService {
     @InjectModel(Entry.name) private entryModel: Model<EntryDocument>,
   ) {}
 
-  async getAll(user: UserDocument): Promise<CategoryDocument[]> {
+  async getAll(ownerId: string): Promise<CategoryDocument[]> {
     return await this.categoryModel
       .find({
-        owner: new mongoose.Types.ObjectId(user._id),
+        owner: new mongoose.Types.ObjectId(ownerId),
       })
       .populate('items');
   }
@@ -32,40 +32,52 @@ export class CategoryService {
   }
 
   async getCategory(id: string, ownerId: string): Promise<CategoryDocument> {
-    return await this.categoryModel
+    const category = await this.categoryModel
       .findOne({
         owner: new mongoose.Types.ObjectId(ownerId),
         _id: new mongoose.Types.ObjectId(id),
       })
       .populate('items');
+    if (!category) {
+      throw new NotFoundException(`Category (id=${id}) was not found`);
+    }
+    return category;
   }
 
-  async deleteCategory(id: string, user: UserDocument): Promise<void> {
+  async deleteCategory(id: string, ownerId: string): Promise<void> {
     const categoryBeforeDeletion = await this.categoryModel.findOneAndDelete({
       _id: new mongoose.Types.ObjectId(id),
-      ownerId: new mongoose.Types.ObjectId(user._id),
+      ownerId: new mongoose.Types.ObjectId(ownerId),
     });
+    if (!categoryBeforeDeletion) {
+      throw new NotFoundException(`Category ${id} was not found`);
+    }
     await this.entryModel.deleteMany({
       category: new mongoose.Types.ObjectId(categoryBeforeDeletion._id),
-      owner: new mongoose.Types.ObjectId(user._id),
+      owner: new mongoose.Types.ObjectId(ownerId),
     });
   }
 
   async editCategory(
     id: string,
-    user: UserDocument,
+    ownerId: string,
     categ: Partial<CategoryDocument>,
   ): Promise<CategoryDocument> {
-    const updated = await this.categoryModel.findOneAndUpdate(
-      {
-        _id: new mongoose.Types.ObjectId(id),
-        owner: new mongoose.Types.ObjectId(user._id),
-      },
-      categ,
-      {
-        returnDocument: 'after',
-      },
-    );
-    return updated.populate('items');
+    const updated = await this.categoryModel
+      .findOneAndUpdate(
+        {
+          _id: new mongoose.Types.ObjectId(id),
+          owner: new mongoose.Types.ObjectId(ownerId),
+        },
+        categ,
+        {
+          returnDocument: 'after',
+        },
+      )
+      .populate('items');
+    if (!updated) {
+      throw new NotFoundException(`Category ${id} was not found`);
+    }
+    return updated;
   }
 }
